Type alias for DependingTransactionsMap.
Introduce and use a type alias for the std::map type that holds the data
for mapping outputs to depending orphan transactions in
BlockMemoryPoolTransactionCollector.

This is quite a length type when spelled out, and used in a couple of
functions as argument.  Using a type alias makes it more readable and
also easier to adjust in the future as needed.