Remove begin_ptr/end_ptr from Logging.cpp.
Logging.cpp defined a "private" version of the begin_ptr and end_ptr
templates that are also defined in serialize.h.  This replaces that
with simply using serialize.h.

In addition to avoiding code duplication, this also makes sure that the
code still works even if Logging.cpp were ever to include directly
or indirectly serialize.h (where without this commit, it would lead
to an ambiguous template call).