Fix out-of-bounds access in std::vector.
The code changed here did access blockUndo.vtxundo out of bounds (at index
-1) for the coinbase transaction of a block (transactionIndex == 0).  This
had no practical implication as the reference passed into
UpdateCoinsReversingTransaction was not used in that situation, but it is
still undefined behaviour, and the fix is easy.

The change also makes it more explicit that txundo could be "invalid" in
UpdateCoinsReversingTransaction, and in which situation it should be
accessed.