Set -nolistenonion in all regtests.
Previously, we set -nolistenonion selectively in some of the regtests.
Without this flag, divid tries to connect to a locally-running Tor instance
and listen with a hidden service on mainnet, which we do not want in tests
and which causes them to use far more CPU and time than without.

In this change, we simply set the flag explicitly in the test framework
for *all* regtests (which is really what we want) instead of setting it
individually in some of them.