Refactor mempool_tests.
This refactors the existing mempool unit tests.  We move common code
(that adds our test transactions to the pool) into a method
in the test fixture.  Also, we set up a proper CBlockIndex and
CCoinsView so that the transactions are "valid" based on them.  This
allows us to do a mempool sanity check as well, extending test coverage
and making the test more robust in general.