Assert that locks in use are not destructed.
This adds (very light-weight) custom logic to CCriticalSection and
related classes that tracks the lock status of each mutex, and
asserts that the mutex is fully unlocked when it gets destructed.

This in particular also ensures that addresses of mutexes cannot be
reused while a lock (referenced by void* address) is still being used
and on the lock stack of any thread.