Define CTxMemPool::lookupOutpoint.
The memory pool mainly tracks transactions by txid; however, in some
places, we actually need to look up the transaction by the output hash
it creates (which will be different from the txid after segwit light).

This introduces a new method CTxMemPool::lookupOutpoint, which is meant
to be used for these situations (even though for now it does the same
as lookup).  It also updates the code to use the new method where
this is appropriate.