Define fork activation for CLTV.
This defines a new Fork::CheckLockTimeVerify activation, which enables
the SCRIPT_VERIFY_CHECKLOCKTIMEVERIFY flag for block validation.  For
the standard script verification flags, we unconditionally enable
the new check.

For now, the fork is scheduled at dummy time 2'000'000'000, which we can
use for testing and the initial merge until the fork can be scheduled.