Define FlatDataFile superclass.
This introduces a new FlatDataFile superclass for CFlatDB, which takes
care of holding the filename/path and also the writing of a chunk of
data in a flexible way.

It is a first step for fully separating chunked reading/writing out of
CFlatDB, and later introducing another subclass for an append-only file.