Correctly handle standardness of P2SH inputs.
The old logic for handling standardness with P2SH inputs has a loophole:
When an input is found which is of unknown type but with an acceptable
number of sigops, the whole transaction is considered fine.  This fixes
the logic to properly check all following inputs as well -- otherwise
this opens up a venue for having inputs that should be considered
non-standard accepted as standard.