Allow CCoinsViewBacked to use a read-only base.
This extends the functionality in CCoinsViewBacked and derived classes,
so that they can be constructed with a read-only (const) base cache.
In this case, most methods are still available, and only the BatchWrite
method will fail when called.

This allows many paths in the code to be modified so that they work
with a const CCoinsView for things that actually only need read access
to the underlying coins view (e.g. DB verification).