Encapsulate construction of shallow databases.
This refactors how the shallow databases are constructed and destructed.
Instead of managing a bunch of pointers directly from init, the instances
needed for that are wrapped up into a single class, and then we use a global
singleton instance of that class.

The external interface is (more or less) then just a pair of Setup and
Cleanup methods.