Make CBlockIndex'es inside CChain const.
Once a CBlockIndex instance has been added to the active chain
(or any CChain for that matter), it has already been fully validated
and is read-only from that point on, except for a few minor exceptions
(like the invalidateblock RPC) and unit tests that manipulate a chain
directly rather than through the production logic.

This change updates CChain to store and expose its blocks as const
exclusively.  This helps enforce this expectation of read-only access
to blocks from the active chain in the code.