Make fJustCheck in ConnectBlock useful.
ConnectBlock has a fJustCheck flag, which was until now actually never
set to true.  This change changes what the flag does:  In the new code,
the flag can be used to control whether or not to update on-disk data
like the block undo and address/spent indices.

This new behaviour is useful to turn off those side effects in VerifyDB,
where we now actually set the flag to true.  Changes in particular to the
coins view are desired in this situation (as we use an in-memory overlay
view there).

The commit also removes some dead logic from ConnectBlock.