Move actual chain state into ChainstateManager.
This moves the actual instances holding the chainstate (like the
CChain for activeChain, or the CCoinsViewCache / CCoinsViewDB
for the UTXO set) into ChainstateManager, rather than having
them as globals and just referring to them from there.

After this change, it gets strictly enforced that the state is only
accessed through ChainstateManager (which is already largely the case),
and also makes it more explicit how the life cycle is managed (namely
through that one instance of ChainstateManager created/destructed
from init.cpp).