Move some code outside of ShallowDatabases.
This moves some code that is actually unrelated to the databases themselves
(for the spork manager and for flushing the state on shutdown) from the
ShallowDatabases setup/cleanup code outside to where the calls happen.

After this change, the ShallowDatabases themselves are really just
initialising and cleaning up the database instances as well as the
global pointers to them.  But no other (in particular, "complex") logic
is happening inside the class.