Non-static helper modules.
At some places, "helper modules" (like the LotteryWinningsCalculator,
BlockIndexLotteryUpdater or ProofOfStakeModule) are declared as
"static globals".

This creates the issue that they are, indeed, globals (even if the
initialisation order is well defined) and don't have a proper life cycle,
and thus those instances survive beyond normal node shutdown (which
means they will contain dangling references to things like the
ChainstateManager).

Also, all of these instances are trivial to construct, as they just
collect a bunch of references to other dependencies together.  Hence,
we can just as well construct (and properly destruct) them on each
call to the relevant part of the code, which gets rid of their
"global" nature and makes sure all instances and life times are
properly synchronised to each other.