Manage CSporkManager through init.
Instead of having a "static global" instance for the CSporkManager
(that is initialised on first use and never destructed), we now
manage the global CSporkManager instance through the logic in init.cpp.

This ensures that it has a well-defined life cycle, getting constructed
after the ChainstateManager and destructed before (at the points in time
where previously the spork database was allocated/deallocated).

The instance is still global (exposed through GetSporkManager() as before),
which is something to work on separately.